--- gemstone/src/swapper/cetus/math/mod.rs
@@ -1,9 +1,0 @@
-pub mod clmm;
-pub mod compute_swap;
-pub mod constants;
-pub mod error;
-pub mod swap;
-pub mod tick;
-pub mod utils;
-
-pub use clmm::{compute_swap, ClmmPoolData, TickData};
